panelize.py: Use float instead of int for coordinates values
KiCad now use metric system and millimeter unit as default. Most parsed
values must be read and written as float to adapt to this change.
This commit reads and writes, in millimeter, as float:
 - Coord() arguments
 - charwidth
 - charheight
 - linewidth
This is not past-proof, as it should be better to read the file Units:
field and adapt the reading/writing accordingly.